refactor(app.js): Extract upload image
